requirements.txt
flask
dash
plotly
pandas
gunicorn
python-dotenv